+++ R/config.R
@@ -2,7 +2,7 @@
 #' @description Compute the internal runtime parameter list of
 #' \code{\link{make}()}. This could save time if you are planning
 #' multiple function calls of functions like \code{\link{outdated}()}
-#' or \code{\link{plot_graph}()}. Drake needs to import and cache files
+#' or \code{\link{vis_drake_graph}()}. Drake needs to import and cache files
 #' and objects to compute the configuration list, which in turn
 #' supports user-side functions to help with visualization and parallelism.
 #' The result differs from
@@ -11,7 +11,8 @@
 #' not just the imports.
 #' @export
 #' @return The master internal configuration list of a project.
-#' @seealso \code{\link{workplan}}, \code{\link{make}}, \code{\link{plot_graph}}
+#' @seealso \code{\link{workplan}}, \code{\link{make}},
+#' \code{\link{vis_drake_graph}}
 #' @param plan same as for \code{\link{make}}
 #' @param targets same as for \code{\link{make}}
 #' @param envir same as for \code{\link{make}}
@@ -52,7 +53,7 @@
 #' # In make(..., jobs = n), it would be silly to set `n` higher than this:
 #' max_useful_jobs(config = con)
 #' # Show a visNetwork graph
-#' plot_graph(config = con)
+#' vis_drake_graph(config = con)
 #' # Get the underlying node/edge data frames of the graph.
 #' dataframes_graph(config = con)
 #' }
@@ -110,11 +111,11 @@ config <- function(
   )
   trigger <- match.arg(arg = trigger, choices = triggers())
   if (is.null(graph)){
-    graph <- build_graph(plan = plan, targets = targets,
+    graph <- build_drake_graph(plan = plan, targets = targets,
       envir = envir, verbose = verbose, jobs = jobs,
       skip_imports = skip_imports)
   } else {
-    graph <- prune_graph(graph = graph, to = targets, jobs = jobs)
+    graph <- prune_drake_graph(graph = graph, to = targets, jobs = jobs)
   }
   config <- list(
     plan = plan, targets = targets, envir = envir, cache = cache,