+++ tests/testthat/test-other-features.R
@@ -98,8 +98,8 @@ test_with_dir("targets can be partially specified", {
   expect_silent(check_plan(pl, verbose = FALSE))
 })
 
-test_with_dir("as_file quotes properly", {
-  expect_equal(as_file("x"), "'x'")
+test_with_dir("as_drake_filename quotes properly", {
+  expect_equal(as_drake_filename("x"), "'x'")
 })
 
 test_with_dir("unique_random_string() works", {